# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/integrations/1c/webhook/{type}": {
      "post": {
        "operationId": "IntegrationController_receive",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Вебхук 1С: POST /integrations/1c/webhook/{тип} (HMAC, идемпотентно)",
        "tags": [
          "Integration - 1C"
        ]
      }
    },
    "/api/v1/integrations/status": {
      "get": {
        "operationId": "IntegrationController_status",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Состояние обмена: настроен ли, что ждёт отправки",
        "tags": [
          "Integration - 1C"
        ]
      }
    },
    "/api/v1/integrations/messages": {
      "get": {
        "operationId": "IntegrationController_messages",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Журнал обмена: исходящие и входящие",
        "tags": [
          "Integration - 1C"
        ]
      }
    },
    "/api/v1/integrations/1c/sync/orders": {
      "post": {
        "operationId": "IntegrationController_syncOrders",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Забрать заказы клиентов из 1С (GET A) — пакетом",
        "tags": [
          "Integration - 1C"
        ]
      }
    },
    "/api/v1/integrations/1c/sync/order/{orderNumber}": {
      "post": {
        "operationId": "IntegrationController_syncOrder",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Забрать один заказ из 1С по номеру",
        "tags": [
          "Integration - 1C"
        ]
      }
    },
    "/api/v1/integrations/1c/sync/procurement/{orderNumber}": {
      "post": {
        "operationId": "IntegrationController_syncProcurement",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Забрать закуп под заказ (GET D → GET C)",
        "tags": [
          "Integration - 1C"
        ]
      }
    },
...
    "/api/v1/integrations/outbox/flush": {
      "post": {
        "operationId": "IntegrationController_flush",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Отправить накопленные исходящие сейчас",
        "tags": [
          "Integration - 1C"
        ]
      }
    },
    "/api/v1/integrations/inbox/process": {
      "post": {
        "operationId": "IntegrationController_process",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Разобрать необработанные входящие",
        "tags": [
          "Integration - 1C"
        ]
      }
    },
    "/api/v1/integrations/messages/{id}/retry": {
      "post": {
        "operationId": "IntegrationController_retry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Повторить доставку сообщения (сброс счётчика попыток)",
        "tags": [
          "Integration - 1C"
        ]
      }
    },
...
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get order by ID",
        "tags": [
          "Orders"
        ]
      },
      "patch": {
        "operationId": "OrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update order (field-level RBAC)",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/v1/min-stock-levels": {
      "get": {
        "operationId": "MinStockController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List min stock levels",
        "tags": [
          "Min Stock Levels"
        ]
      }
    },
    "/api/v1/min-stock-levels/{articleId}": {
      "patch": {
        "operationId": "MinStockController_update",
        "parameters": [
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update min stock level",
        "tags": [
          "Min Stock Levels"
        ]
      }
    },
...
  },
  "info": {
    "title": "ЦМК АВРОРА — ERP REST API",
    "description": "REST API v1 ЦМК АВРОРА — планирование производства, заказы, склад, финансы",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}